+++ docs/examples/buttonsocial/facebook.tsx
@@ -1,0 +1,9 @@
+import { ButtonSocial, Flex } from 'gestalt';
+
+export default function Example() {
+  return (
+    <Flex alignItems="center" height="100%" justifyContent="center" width="100%">
+      <ButtonSocial service="facebook" type="continue" />
+    </Flex>
+  );
+}