+++ packages/gestalt/src/index.ts
@@ -13,6 +13,7 @@ import Box from './Box';
 import Button from './Button';
 import ButtonGroup from './ButtonGroup';
 import ButtonLink from './ButtonLink';
+import ButtonSocial from './ButtonSocial';
 import ButtonToggle from './ButtonToggle';
 import Checkbox from './Checkbox';
 import Collage from './Collage';
@@ -110,6 +111,7 @@ export {
   Button,
   ButtonGroup,
   ButtonLink,
+  ButtonSocial,
   ButtonToggle,
   Checkbox,
   Collage,